Port shop screen 1:1 from XML
Part one of fixing the shop screen. For some reason in owo-lib 1.21 scroll components can prevent slots from being clickable.
This was not an issue in 1.20, and therefore it needs special handling to be fixed.
Moving the UI to code seems easier, since some part of the screen needs to be rewritten.
Unfortunately more work needs to be done, due to the nature of XML components naturally stacking, vs. code components naturally respecting the flow layout (being place side-by-side)